refactor response class
