+++ docs/source/api/core.rst
@@ -41,15 +41,6 @@ Keys
 
 .. autofunction:: get_data_keys
 
-Server Response
----------------
-
-.. module:: fseconomy.core.response
-
-.. autoclass:: Response
-   :members:
-   :undoc-members:
-
 Data Decoders
 -------------
