+++ docs/source/api/response.rst
@@ -1,0 +1,10 @@
+.. _response:
+
+Server Response
+===============
+
+.. module:: fseconomy.response
+
+.. autoclass:: Response
+   :members:
+   :undoc-members:
